feat: TExtractQuery
